fix: show correct net/gross prices for order again
