+++ apps/web/app/components/OrderAgainItem/types.ts
@@ -1,0 +1,6 @@
+import type { OrderItem, Order } from '@plentymarkets/shop-api';
+
+export type OrderAgainItemProps = {
+  item: OrderItem;
+  order: Order;
+};
