+++ package.json
@@ -40,7 +40,7 @@
   },
   "dependencies": {
     "@plentymarkets/shop-core": "^1.10.0",
-    "@plentymarkets/shop-api": "^0.137.0",
+    "@plentymarkets/shop-api": "^0.139.0",
     "@plentymarkets/shop-module-gtag": "^1.1.2",
     "@plentymarkets/shop-module-mollie": "^1.2.0",
     "@types/applepayjs": "^14.0.8",
